Moved the following private methods from CustomerSubscriptionController to Errors module: check_params, check_customer_exists, check_subscription_exists, and check_customer_subscription_exists.
