+++ app/controllers/api/v1/customer_subscription_controller.rb
@@ -1,6 +1,10 @@
 class Api::V1::CustomerSubscriptionController < ApplicationController
+  include Errors
   before_action :check_customer_subscription_exists, only: [:update]
   before_action :check_params, :check_customer_exists, :check_subscription_exists, except: [:show]
+  # :check_params
+
+  # require "pry"; binding.pry
 
   def show
     customer = Customer.find(params[:id])
@@ -13,6 +17,8 @@ class Api::V1::CustomerSubscriptionController < ApplicationController
   end
 
   def create
+    # require 'errors'
+    # require "pry"; binding.pry
     new_customer_subscription = CustomerSubscription.create(customer_subscription)
 
     if new_customer_subscription.save
@@ -32,27 +38,27 @@ class Api::V1::CustomerSubscriptionController < ApplicationController
     params.require(:customer_subscription).permit(:customer_id, :subscription_id)
   end
 
-  def check_params
-    if !customer_subscription[:customer_id].present? || !customer_subscription[:subscription_id].present?
-      render json: {error: "Please provide both a customer id and a subscription id."}, status: 400
-    end
-  end
+  # def check_params
+  #   if !customer_subscription[:customer_id].present? || !customer_subscription[:subscription_id].present?
+  #     render json: {error: "Please provide both a customer id and a subscription id."}, status: 400
+  #   end
+  # end
 
-  def check_customer_exists
-    if !Customer.exists?(customer_subscription[:customer_id])
-      render json: {error: "The customer does not exist."}, status: 400
-    end
-  end
+  # def check_customer_exists
+  #   if !Customer.exists?(customer_subscription[:customer_id])
+  #     render json: {error: "The customer does not exist."}, status: 400
+  #   end
+  # end
 
-  def check_subscription_exists
-    if !Subscription.exists?(customer_subscription[:subscription_id])
-      render json: {error: "The subscription does not exist."}, status: 400
-    end
-  end
+  # def check_subscription_exists
+  #   if !Subscription.exists?(customer_subscription[:subscription_id])
+  #     render json: {error: "The subscription does not exist."}, status: 400
+  #   end
+  # end
 
-  def check_customer_subscription_exists
-    if !CustomerSubscription.exists?(customer_id: customer_subscription[:customer_id], subscription_id: customer_subscription[:subscription_id])
-      render json: {error: "The customer does not have this subscription."}, status: 400
-    end
-  end
+  # def check_customer_subscription_exists
+  #   if !CustomerSubscription.exists?(customer_id: customer_subscription[:customer_id], subscription_id: customer_subscription[:subscription_id])
+  #     render json: {error: "The customer does not have this subscription."}, status: 400
+  #   end
+  # end
 end
