+++ config/application.rb
@@ -19,6 +19,7 @@ Bundler.require(*Rails.groups)
 
 module TeaSubscription
   class Application < Rails::Application
+    config.eager_load_paths += %W(#{config.root}/lib) #newly added
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 5.2
