+++ spec/requests/api/v1/customer_subscriptions_index_spec.rb
@@ -61,8 +61,6 @@ describe "Show all a customer's tea subscriptions" do
 
       customer_subscriptions = JSON.parse(response.body, symbolize_names: true)
 
-      require "pry"; binding.pry
-
       expect(customer_subscriptions).to be_a(Hash)
       expect(customer_subscriptions.keys).to match_array([:data])
       expect(customer_subscriptions[:data]).to be_a(Array)
